rename to drawer and add functional tabs + error handling
