+++ ui/src/lib/components/index.ts
@@ -2,5 +2,5 @@
 // SPDX-FileCopyrightText: 2024-Present The UDS Authors
 
 export { default as AnsiDisplay } from './AnsiDisplay/component.svelte'
-export { default as DataTable } from './DataTable/component.svelte'
-export { default as SidePanel } from './SidePanel/component.svelte'
+export { default as DataTable } from './k8s/DataTable/component.svelte'
+export { default as Drawer } from './k8s/Drawer/component.svelte'
