+++ ui/src/lib/components/k8s/Drawer/styles.postcss
@@ -1,0 +1,8 @@
+#drawer-tabs {
+  button {
+    @apply border-transparent rounded-t-lg hover:text-gray-300 hover:border-gray-300 inline-block p-4;
+  }
+  button.active {
+    @apply text-blue-500 border-b-2 border-blue-500;
+  }
+}
