+++ ui/src/lib/features/toast/store.ts
@@ -13,6 +13,8 @@ export type Toast = {
 export const toast = writable<Toast[]>([])
 
 export const addToast = (newToast: Toast) => {
+  console.log('addToast', newToast)
+
   toast.update((toasts) => {
     const id = Date.now() + Math.random()
     const toast = { id, ...newToast }
